requirements.txt
aiosqlite==0.21.0
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
mysql-connector-python==9.3.0
packaging==25.0
parameterized==0.9.0
PyJWT==2.10.1
python-environ==0.4.54
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
typing==3.7.4.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
